requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Flask==2.1.2
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
requests==2.28.0
smmap==5.0.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
